+++ src/Backend/Modules/FormBuilder/Js/FormBuilder.js
@@ -30,6 +30,7 @@ jsBackend.FormBuilder = {
       emptyMessage: jsBackend.locale.msg('NoEmailaddresses'),
       addLabel: utils.string.ucfirst(jsBackend.locale.lbl('Add', 'Core')),
       removeLabel: utils.string.ucfirst(jsBackend.locale.lbl('Delete')),
+      errorMessage: utils.string.ucfirst(jsBackend.locale.err('AddTextBeforeSubmitting')),
       canAddNew: true
     })
   },
@@ -206,7 +207,10 @@ jsBackend.FormBuilder.Fields = {
             case 'radiobuttonDialog':
               jsBackend.FormBuilder.Fields.saveRadiobutton()
               break
-            case 'checkboxDialog':
+            case 'mailmotorDialog':
+              jsBackend.FormBuilder.Fields.saveMailmotorbutton()
+              break
+            case 'checkboxDialogcheckboxDialog':
               jsBackend.FormBuilder.Fields.saveCheckbox()
               break
           }
@@ -218,6 +222,7 @@ jsBackend.FormBuilder.Fields = {
             $('input#dropdownValues').multipleTextbox({
               splitChar: '|',
               emptyMessage: jsBackend.locale.msg('NoValues'),
+              errorMessage: utils.string.ucfirst(jsBackend.locale.err('AddTextBeforeSubmitting')),
               addLabel: utils.string.ucfirst(jsBackend.locale.lbl('Add')),
               removeLabel: utils.string.ucfirst(jsBackend.locale.lbl('Delete')),
               showIconOnly: true,
@@ -229,6 +234,7 @@ jsBackend.FormBuilder.Fields = {
               emptyMessage: jsBackend.locale.msg('NoValues'),
               addLabel: utils.string.ucfirst(jsBackend.locale.lbl('Add')),
               removeLabel: utils.string.ucfirst(jsBackend.locale.lbl('Delete')),
+              errorMessage: utils.string.ucfirst(jsBackend.locale.err('AddTextBeforeSubmitting')),
               showIconOnly: true,
               afterBuild: jsBackend.FormBuilder.Fields.multipleTextboxCallback
             })
@@ -238,6 +244,7 @@ jsBackend.FormBuilder.Fields = {
               emptyMessage: jsBackend.locale.msg('NoValues'),
               addLabel: utils.string.ucfirst(jsBackend.locale.lbl('Add')),
               removeLabel: utils.string.ucfirst(jsBackend.locale.lbl('Delete')),
+              errorMessage: utils.string.ucfirst(jsBackend.locale.err('AddTextBeforeSubmitting')),
               showIconOnly: true,
               afterBuild: jsBackend.FormBuilder.Fields.multipleTextboxCallback
             })
@@ -413,10 +420,16 @@ jsBackend.FormBuilder.Fields = {
                 $('#textboxValue').val(utils.string.htmlDecode(data.data.field.settings.default_values))
                 $('#textboxPlaceholder').val(utils.string.htmlDecode(data.data.field.settings.placeholder))
                 $('#textboxClassname').val(utils.string.htmlDecode(data.data.field.settings.classname))
+                $('#textboxAutocomplete').val(utils.string.htmlDecode(data.data.field.settings.autocomplete))
                 if (data.data.field.settings.reply_to &&
                   data.data.field.settings.reply_to === true
                 ) {
                   $('#textboxReplyTo').prop('checked', true)
+                }
+                if (data.data.field.settings.use_to_subscribe_with_mailmotor &&
+                  data.data.field.settings.use_to_subscribe_with_mailmotor === true
+                ) {
+                  $('#textboxMailmotor').prop('checked', true)
                 }
                 if (data.data.field.settings.send_confirmation_mail_to &&
                   data.data.field.settings.send_confirmation_mail_to === true
@@ -479,6 +492,7 @@ jsBackend.FormBuilder.Fields = {
                 $('#datetimeValueType').val(utils.string.htmlDecode(data.data.field.settings.value_type))
                 $('#datetimeType').val(utils.string.htmlDecode(data.data.field.settings.input_type))
                 $('#datetimeClassname').val(utils.string.htmlDecode(data.data.field.settings.classname))
+                $('#datetimeAutocomplete').val(utils.string.htmlDecode(data.data.field.settings.autocomplete));
                 $.each(
                   data.data.field.validations,
                   function (k, v) {
@@ -603,6 +617,15 @@ jsBackend.FormBuilder.Fields = {
 
                 // show dialog
                 $('#checkboxDialog').modal('show')
+              } else if (data.data.field.type === 'mailmotor') {
+                // mailmotor edit
+                // fill in form
+                $('#mailmotorId').val(data.data.field.id)
+                $('#mailmotorLabel').val(utils.string.htmlDecode(data.data.field.settings.label))
+                $('#mailmotorListId').val(utils.string.htmlDecode(data.data.field.settings.list_id))
+
+                // show dialog
+                $('#mailmotorDialog').modal('show')
               } else if (data.data.field.type === 'heading') {
                 // heading edit
                 // fill in form
@@ -841,6 +864,7 @@ jsBackend.FormBuilder.Fields = {
     var validation = $('#datetimeValidation').val()
     var errorMessage = $('#datetimeErrorMessage').val()
     var classname = $('#datetimeClassname').val()
+    var autocomplete = $('#datetimeAutocomplete').val()
 
     // make the call
     $.ajax(
@@ -857,7 +881,8 @@ jsBackend.FormBuilder.Fields = {
           input_type: inputType,
           validation: validation,
           error_message: errorMessage,
-          classname: classname
+          classname: classname,
+          autocomplete: autocomplete
         }),
         success: function (data, textStatus) {
           // success
@@ -1156,6 +1181,65 @@ jsBackend.FormBuilder.Fields = {
   },
 
   /**
+   * Handle mailmotorButton save
+   */
+  saveMailmotorbutton: function () {
+    // init vars
+    var fieldId = $('#mailmotorId').val()
+    var type = 'mailmotor'
+    var label = $('#mailmotorLabel').val()
+    var listId = $('#mailmotorListId').val()
+
+    // make the call
+    $.ajax({
+      data: $.extend({}, jsBackend.FormBuilder.Fields.paramsSave, {
+        form_id: jsBackend.FormBuilder.formId,
+        field_id: fieldId,
+        type: type,
+        label: label,
+        list_id: listId
+      }),
+      success: function (data, textStatus) {
+        // success
+        if (data.code === 200) {
+          // clear errors
+          $('.jsFieldError').html('')
+
+          // form contains errors
+          if (typeof data.data.errors !== 'undefined') {
+            // assign errors
+            if (typeof data.data.errors.label !== 'undefined') {
+              $('#mailmotorLabelError').html(data.data.errors.label)
+            }
+
+            if (typeof data.data.errors.list_id !== 'undefined') {
+              $('#mailmotorListIdError').html(data.data.errors.list_id)
+            }
+
+            // toggle error messages
+            jsBackend.FormBuilder.Fields.toggleValidationErrors('mailmotorDialog')
+          } else {
+            // saved!
+            // append field html
+            jsBackend.FormBuilder.Fields.setField(data.data.field_id, data.data.field_html)
+
+            // close console box
+            $('#mailmotorDialog').modal('hide')
+          }
+        } else {
+          // show error message
+          jsBackend.messages.add('danger', textStatus)
+        }
+
+        // alert the user
+        if (data.code !== 200 && jsBackend.debug) {
+          window.alert(data.message)
+        }
+      }
+    })
+  },
+
+  /**
    * Handle submit save
    */
   saveSubmit: function () {
@@ -1293,6 +1377,9 @@ jsBackend.FormBuilder.Fields = {
             if (typeof data.data.errors.reply_to_error_message !== 'undefined') {
               $('#textboxReplyToErrorMessageError').html(data.data.errors.reply_to_error_message)
             }
+            if (typeof data.data.errors.use_to_subscribe_with_mailmotor_error_message !== 'undefined') {
+              $('#textboxMailmotorErrorMessageError').html(data.data.errors.use_to_subscribe_with_mailmotor_error_message)
+            }
 
             // toggle error messages
             jsBackend.FormBuilder.Fields.toggleValidationErrors('textareaDialog')
@@ -1336,6 +1423,8 @@ jsBackend.FormBuilder.Fields = {
     var validationParameter = $('#textboxValidationParameter').val()
     var errorMessage = $('#textboxErrorMessage').val()
     var classname = $('#textboxClassname').val()
+    var mailmotor = $('#textboxMailmotor').is(':checked')
+    var autocomplete = $('#textboxAutocomplete').val()
 
     // make the call
     $.ajax({
@@ -1346,6 +1435,7 @@ jsBackend.FormBuilder.Fields = {
         label: label,
         default_values: value,
         reply_to: replyTo,
+        use_to_subscribe_with_mailmotor: mailmotor,
         send_confirmation_mail_to: sendConfirmationMailTo,
         confirmation_mail_subject: confirmationMailSubject,
         required: required,
@@ -1354,7 +1444,8 @@ jsBackend.FormBuilder.Fields = {
         validation_parameter: validationParameter,
         error_message: errorMessage,
         placeholder: placeholder,
-        classname: classname
+        classname: classname,
+        autocomplete: autocomplete
       }),
       success: function (data, textStatus) {
         // success
@@ -1379,6 +1470,9 @@ jsBackend.FormBuilder.Fields = {
             }
             if (typeof data.data.errors.reply_to_error_message !== 'undefined') {
               $('#textboxReplyToErrorMessageError').html(data.data.errors.reply_to_error_message)
+            }
+            if (typeof data.data.errors.use_to_subscribe_with_mailmotor_error_message !== 'undefined') {
+              $('#textboxMailmotorErrorMessageError').html(data.data.errors.use_to_subscribe_with_mailmotor_error_message)
             }
             if (typeof data.data.errors.send_confirmation_mail_to_error_message !== 'undefined') {
               $('#textboxSendConfirmationMailToErrorMessageError').html(data.data.errors.send_confirmation_mail_to_error_message)